feat: standalone binary
In seeing some people attempt to use the library, I realized it may be
better to ship it as a standalone binary:

- The class it generates has no dependencies.
- Generally, you will generate the preload file once, and then
  manipulate it to provide additional rules. Once generated, you no
  longer need the tooling

This patch provides a `bin/phly-opcache-preload` file exposed as a
vendor binary in order to allow global installation. When used in this
way, the commands become `generate:preload-file` and `generate:ini`.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>